Add Role-Based Credentials Support
-Role-Based chain has the highest precedence, but requires first the IAM
user credentials to be able to get the role credentials. So first the
regular provider chain is built and then supplied to the role-based
providers in their own chain. Finally they are joined together for the
final complete credential provider chain
-Bump minor release version to 0.8.0-SNAPSHOT
-Supports all the provider modes of the standard credentials: specify
the roleArn in systemProperty, env variable, or property files.